Switch the authentication to be OIDC based (#278)
* Add new dependencies

* Fix vite setup

* Create an OpenIdService

* Implement getRedirectUrls
This using rxjs to build up a cached pipeline
that uses OIDC to discover the correct URLs,
then generates the correct client-side secrets
(such as the code challenge and verifier),
and then returns the URL's

This is sufficient for a frontend to be able to redirect to an OIDC
provider in order to initiate an auth flow

* Implement the LoginPrompt

* Handle the redirect

* Persist the access_token
This changes the openid service to have a getAccessToken observable
which reads from localstorage and keeps it in sync.
Next, an AuthProvider and corresponding useAuth hook is created
in order to be able to read this data and bridge it to react
Finally, all cases of useGoogleAuth were replaced with this new
functionality

* Remove google-auth.ts

* Refresh the token when it is close to expiring

* Remove all the old backend auth code

* Not-working backend authentication using OAuth
Anaconda returns a 403 for the userinfo endpoint

* Grab the user claims and store them alongside the access token
Next, use the real user info instead of the dummy data

* Run prettier

* Fix environment variables for CI

* PR feedback

* Simplify the redirect handler

* DRY unit tests

* Fix local-developer mode

* Add getUser and getAccessToken to the AuthProvider

* Remove useCurrentUserId

* Remove useCurrentUser
The user syncing code is moved into the UserProfile instead
Other complexities are removed, because there are no anonymous
users right now

* Fix merge conflict

* Change the backend validation code to directly verify the JWT using JWKS

* Fix the hardcoded algorithm check for local testing

* Update documentation

* Run the formatter

* Allow processing both AUTH_TOKEN and oauth tokens in the backend

* PR feedback and bug fixes

* Set environment variables - note are secrets

* Access token bug fixes
First, we add a singleton around getTokens, to ensure there is only
one active observable. This prevents multiple token refreshes
from happening at the same time, which can cause some streams to error out
and potentially log the user out.

Secondly, we create a new "freshness" observable to ensure that the
access token is updated when it nears expiry

* Change getUser and getAccessToken to be getters
This allows for destructuring on the call-side and removes the need for the function call

* update client IDs

---------

Co-authored-by: Kyle Kelley <rgbkrk@gmail.com>