Switch all of the backend/ files to use workerGlobals The global types in cloudflare are different than the ones in node. This includes things like Request, Response, and more
The old code is broken because it imports types for the browser, node, and cloudflare,
creating a kind of merged environment. This is broken in ways that are both obvious and subtle.

The short-term fix is to be extremely specific about cloudflare types. No more using the global
Request object. Instead, everything needs to go through the backend/types file, where
both the types and the global variables are defined and used.

Since this fixes the remaining linter errors for test files, this change also enforces
the test files pass typescript on the CI