Add user saved prompt (#582)
* AI: Add system prompt

---

@index.ts @AiPanel.tsx @migrations/ @db.ts

Add a trpc endpoint to create or modify the system prompt for the user. Put the db code in the db file. We could potentially have more than one per user, so create a table for system prompts with columns:
- user id
- system prompt
- ai model

for AI model, for now default to null

In the UI, alway assume one system prompt for now

* Set notebook metadata for system prompt from user

* Extract system prompt editor

* Allow resize

* Clean up mutation logic

* Tweak UX and logic of system prompt editing

* Fix out-of-sync system prompt

* Delete empty file

* Update migration date

* Update error messages

* Clean up upsert to use DB timestamp, return real db values, show toast

* Save on cmd-enter

* Rename system prompt to saved prompt

* Update more

* Use feature flag to show UI

* Update local state when saved prompt changes

* Check for saved prompt in executable cells

* Dark kbd

* Fix accidentally removed ff

* Fix missing hook dep